Fixed sonar.

+++ src/main/java/g2301_2400/s2382_maximum_segment_sum_after_removals/Solution.java
@@ -21,7 +21,7 @@ public class Solution {
                 return;
             }
             this.root[x] = x;
-            this.sum[x] = (long) value;
+            this.sum[x] = value;
         }
 
         public int find(int x) {
